+++ code/espurna/config/hardware.h
@@ -228,13 +228,13 @@
     #define LED1_PIN_INVERSE    0
 
 // -----------------------------------------------------------------------------
-// JanGoe Wifi Relay 
+// Jan Goedeke Wifi Relay
 // https://github.com/JanGoe/esp8266-wifi-relay
 // -----------------------------------------------------------------------------
 
 #elif defined(WIFI_RELAY_NC)
 
-    #define MANUFACTURER        "JANGOE"
+    #define MANUFACTURER        "JAN_GOEDEKE"
     #define DEVICE              "WIFI_RELAY_NC"
     #define BUTTON1_PIN         12
     #define BUTTON1_RELAY       1
@@ -249,7 +249,7 @@
 
 #elif defined(WIFI_RELAY_NO)
 
-    #define MANUFACTURER        "JANGOE"
+    #define MANUFACTURER        "JAN_GOEDEKE"
     #define DEVICE              "WIFI_RELAY_NO"
     #define BUTTON1_PIN         12
     #define BUTTON1_RELAY       1